# vulkan: get the first command buffer submitted sooner